% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_helpers.R
\name{adaptiveCenterCaption}
\alias{adaptiveCenterCaption}
\title{Adds a centered caption to a variety of output types in Rmarkdown.}
\usage{
adaptiveCenterCaption(
  chunckName,
  caption = "",
  capFxn,
  type = knitr::opts_knit$get("rmarkdown.pandoc.to")
)
}
\arguments{
\item{chunckName}{character string. Rmarkdown code chunk name containing the
graphic to label. Passed to the \code{name} option in the function created
by \code{captioner::captioner()}.}

\item{caption}{character string. Caption body to add to graphic. Passed to
the \code{caption} option in the function created by
\code{captioner::captioner()}.}

\item{capFxn}{character string. Name of the function created by
\code{captioner::captioner()} for that graphic type.}

\item{type}{character string. Type of output from the Rmarkdown knit. Likely
not to be changed. Default is set to the result of
\code{knitr::opts_knit$get("rmarkdown.pandoc.to")}}
}
\value{
Returns the plain text markdown code needed to create the centered
  caption for the specified output type.
}
\description{
Adds a centered caption for latex pdf, html, and word (only left
  align works at this time) documents created using Rmarkdown.
  \code{adaptiveCenterCaption()} builds off of the function created using
  \code{captioner::captioner()}, assesses which type of output document is
  being called using \code{knitr::opts_knit$get("rmarkdown.pandoc.to")}, and
  then outputs plain markdown formating text for the chosen document type.
}
\section{Creation notes}{
 First created on 2019-Mar-26 as I was exploring how
  to use Rmarkdown.
}

\examples{
\donttest{
# Define captioner function
```{r, include=FALSE}
figNum <- captioner::captioner(prefix="Fig.")
```

# Plot the figure with center alignment
```{r Plot-Caret-Price, echo=FALSE, results='hide', fig.align='center'}
qplot(carat, price, data = diamonds)
```

# Run the adaptiveCenterCaption function making sure to include results="asis" in chunk options
```{r, echo=FALSE, results="asis"}
adaptiveCenterCaption(chunckName="Plot-Caret-Price", caption = "This is caption 1", capFxn = "figNum")
```
}

}
