% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmHelpers.R
\name{chat}
\alias{chat}
\title{Find overdispersion parameter}
\usage{
chat(object, output = TRUE)
}
\arguments{
\item{object}{Model object that your can be used with \code{residuals()} and
contains the vector \code{$df.residual}. See \code{Details}.}

\item{output}{logical of whether to print output to the console}
}
\value{
Returns a vector with of size 2 with the \eqn{\hat{c}} value and the
associated p-value from the \eqn{\chi^2} testing that the null that
\eqn{\hat{c}} = 1.
}
\description{
\code{chat()} calculates the overdispersion parameter
\eqn{\hat{c}}.
}
\details{
I know that this can work on \code{glm}, \code{MASS::glm.nb}, and
\code{pscl::zeroinfl} models.
}
\section{Creation notes}{
 First created on 2019-Nov-14 in Homework_4.Rmd while
working on homework for FNR647.
}

\examples{

}
