% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleUnitConvert.R
\name{deg2rad}
\alias{deg2rad}
\alias{rad2deg}
\title{Convert Degrees to Radians and Vice Versa}
\usage{
deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{deg}{Input degree you want to change to radians.}

\item{rad}{Input radian you want to change to degrees.}
}
\description{
Base R doesn't have a native function to change degrees to but
  requires radians for trig functions. These are to help handle that.
}
\section{Creation notes}{
 First created in 2019-Feb in FIAspatialAdditions.R
  for my BigDataFNR class presentation on working with spatial data in R in
  my purdueResearch repo.
}

