% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_helpers.R
\name{latex_character}
\alias{latex_character}
\title{Force Latex Character Vector in Kable}
\usage{
latex_character(vec)
}
\arguments{
\item{vec}{character vector that you want to force to output as character in
kable.}
}
\value{
Returns character vector with special latex formating.
}
\description{
\code{latex_character()} forces a row to output in kable as a
character. It is helpful when the vector contains a latex special character
like `_`.
}
\section{Creation notes}{
 First created on 2019-Nov-14 in Homework_4.Rmd while
working on homework for FNR647.
}

\examples{
\donttest{

modelCoefTable(Q1_model, overdispersed = FALSE) %>%
 mutate(Coefficient = latex_character(Coefficient),
        `Pr(>|z|)`= latex_niceSigFig(`Pr(>|z|)`,digits = 3)) %>%
 kable(escape = F, booktabs = T, digits = 3)
}
}
