% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_helpers.R
\name{latex_niceSigFig}
\alias{latex_niceSigFig}
\title{Force Pretty Latex Significant Figure in Kable}
\usage{
latex_niceSigFig(vec, digits = 3)
}
\arguments{
\item{vec}{vector that you want to force to output as sig fig in kable.}

\item{digits}{integer that defines how many sig figs you wnat to keep.}
}
\value{
Returns character vector with special latex formating.
}
\description{
\code{latex_niceSigFig()} formats character vector in data to
output as pretty sig fig
}
\section{Creation notes}{
 First created on 2019-Nov-14 in Homework_4.Rmd while
working on homework for FNR647.
}

\examples{
\donttest{

modelCoefTable(Q1_model, overdispersed = FALSE) %>%
 mutate(Coefficient = latex_character(Coefficient),
        `Pr(>|z|)`= latex_niceSigFig(`Pr(>|z|)`,digits = 3)) %>%
 kable(escape = F, booktabs = T, digits = 3)
}
}
