% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmHelpers.R
\name{modelCoefTable}
\alias{modelCoefTable}
\title{Model Coefficient Table}
\usage{
modelCoefTable(model, overdispersed = FALSE)
}
\arguments{
\item{model}{Model object. See \code{Details}.}

\item{oversdispersed}{logical of whether to recalculate standard error using
the overdispersion parameter \code{\link{chat}}.}
}
\value{
Returns a tibble of coefficients, standard errors, Wald's z, and an
associated p-value.
}
\description{
\code{modelCoefTable()} creates a tibble of coefficients,
standard errors, Wald's z, and an associated p-value.
}
\details{
I know that this can work on \code{glm}, \code{MASS::glm.nb}, and
\code{pscl::zeroinfl} models.
}
\section{Creation notes}{
 First created on 2019-Nov-14 in Homework_4.Rmd while
working on homework for FNR647.
}

\examples{

}
