% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Integer numbers to words}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{Input integer.}
}
\value{
Returns a character string.
}
\description{
\code{numbers2words()} takes an integer argument and outputs
it in word form.
}
\details{
See description.
}
\section{Creation notes}{
 First added on 2019-Sept-7 while trying to finish
the first homework assignment for FNR647. I did not create the underlying
code. This credit goes as follows:

I got this from https://gist.github.com/psychemedia/150cb9901529da58124a and
 added to my helper package because I want it it be easier for me to access
 without sourcing a script.

The author AJH edited from:
 https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r

Who accredits John Fox as the original author
 Function by John Fox found here:
 http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html
}

