% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_Helpers.R
\name{sf_LatLongGrid}
\alias{sf_LatLongGrid}
\title{Make polygon grid over \code{sf} feature using lat-lon coordinates}
\usage{
sf_LatLongGrid(sfLayer, gridSize = 1, prettyGrid = TRUE, CRS = 4269,
  output = c("polygon", "raster"))
}
\arguments{
\item{sfLayer}{sf object}

\item{gridSize}{grid resolution, uses coordinate system measurment default}

\item{prettyGrid}{logical, Use pretty grid measuremetns? Defalut: TRUE}

\item{CRS}{a CRS string or EPSG key}

\item{output}{character, Do you want a polygon or raster output of the
estimated grid?}
}
\value{
Returns either a raster or sf Polygon grid to overlay over the
  sfLayer data.
}
\description{
Make polygon grid over \code{sf} feature using lat-lon
  coordinates. This can be used to split data into regions determined by the
  exported grid. Can be used to create a standard grid using the bounding box
  of the original feature or a 'pretty' grid that locks grid edges to more
  even numbers.
}
\section{Creation notes}{
 First created on 2019-Apr-02 in
  1-2_FIA_CleanRawData.R
}

\examples{

}
