% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_Helpers.R
\name{sf_sample}
\alias{sf_sample}
\title{Better random point sample}
\usage{
sf_sample(num, sampleArea)
}
\arguments{
\item{num}{Integer. How many samples to draw?}

\item{sampleArea}{Polygon depicting study area to draw from.}
}
\value{
Returns an sf object with one column (id) that gives a unique
  identifier to each point created.
}
\description{
\code{sf_sample()} is a wrapper on st_sample that allows it to
  return an sf object with a data frame.
}
\section{Creation notes}{
 First created on 2019-Mar-6 in Lab4inR.R as I was
  working to create a NetLogo Program in R.
}

\examples{

}
