% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_Helpers.R
\name{sf_studyArea}
\alias{sf_studyArea}
\title{Make Study Area Polygons}
\usage{
sf_studyArea(sfLayer, ratio = 1, dist = NULL)
}
\arguments{
\item{sfLayer}{sf object}

\item{ratio}{Numerical vector of length 1 or 2 (as it pertains to the x,y
axes). Default: 1 to create a basic study area polygon for the sfLayer.
Ignored if a \code{dist} is given.}

\item{dist}{Numerical vector of length 1 or 2 (as it pertains to the x,y
axes). Default: \code{NULL}.}
}
\value{
Returns a polygon sf layer that is a rectangle.
}
\description{
Make study area polygons from the extent (bbox) of an sf layer.
  Includes scaling options to increase or decrease size of the polygon by a
  ratio or by a distance calculated in map units. Because of this, the
  distance parameter only works for projected layers.
}
\section{Creation notes}{
 First created on 2019-Apr-22 while working on my IBM
  class project.
}

\section{Future directions}{
 I may need to add checks to make sure negative
  distance values won't flip the min and max values in the extent creation
  section.
}

\examples{

}
