% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_Helpers.R
\name{sf_XYtoCols}
\alias{sf_XYtoCols}
\title{Add XY columns from point sf}
\usage{
sf_XYtoCols(x, names = c("x", "y"))
}
\arguments{
\item{x}{sf Point feature}

\item{names}{Names for x and y columns. (Default is c("x", "y"))}
}
\value{
Returns the inputed sf object with two new columns.
}
\description{
\code{sf_XYtoCols()} adds and x and y column to the sf dataframe
from the sf geometry.
}
\section{Creation notes}{
 First conceptualized on 2019-Mar-6 in Lab4inR.R as
I was working to create a NetLogo Program in R. But I was able to find code
online that fit my needs here: https://github.com/r-spatial/sf/issues/231
}

\examples{

}
