% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterWindowExtract.R
\name{windowExtract}
\alias{windowExtract}
\title{Extract raster values in a window}
\usage{
windowExtract(xf, yf, radius = 1, xydata)
}
\arguments{
\item{xf, yf}{Input cellIDs that you are extracting from.}

\item{radius}{numeric. radius used to define extractWindow}

\item{xydata}{data.frame object that contains the correct headers ('x', 'y',
(valcolumns)). Will need to seperate out this function and create a class
for it.}
}
\value{
Returns a data.frame with the raster values extracted at the current
  radius.
}
\description{
Does the work to extract rasters
}
\section{Creation notes}{
 First created on 2019-Mar-6 in the Lab4inR.R and
  select.window.R scripts in my IBM class folder of the purdueResearch git.
  They were developed using base code from \code{CommEcol::select.window()}.
}

